clean marycontroller + mary startup launcher
