+++ src/main/java/marytts/http/models/constants/ResponseType.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package marytts.http.response;
+package marytts.http.models.constants;
 
 /**
  * Mary response type
